// src/constants/colors.d.ts
export declare const COLORS: {
    white: string;
    green: string;
    purple: string;
    lightPurple: string;
    gray: string;
    lightRed: string;
    red: string;
    text: {
        border: string;
        extraLight: string;
        light: string;
        medium: string;
        dark: string;
        white: string;
        black: string;
        schemesSecondary: string;
    };
    main: {
        dark: string;
        light: string;
        shadowBox: string;
    };
    status: {
        completed: {
            bg: string;
            text: string;
        };
        failed: {
            bg: string;
            text: string;
        };
        notArrived: {
            bg: string;
            text: string;
        };
        atRisk: {
            bg: string;
            text: string;
        };
        upcoming: {
            bg: string;
            text: string;
        };
    };
    background: {
        logoBlock: string;
        lightPurple: string;
    };
};